+++ packages/sync/CHANGELOG.md
@@ -1,6 +1,15 @@
-# 3.0.0
+# Change Log
+
+## 4.0.0
+
+* Updated to use `angel3` packages
+* Published with `angel3` prefix
+
+## 3.0.0
+
 * Migrated to support Dart SDK 2.12.x NNBD
 
-# 2.0.0
+## 2.0.0
+
 * Dart 2 + Angel 2 updates.
 * Extend `StreamChannel`, instead of the defunct `WebSocketSynchronizer`.